+++ tech1-framework-domain/src/main/java/io/tech1/framework/domain/properties/configs/ServerConfigs.java
@@ -1,0 +1,22 @@
+package io.tech1.framework.domain.properties.configs;
+
+import io.tech1.framework.domain.properties.annotations.MandatoryProperty;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+
+// Lombok (property-based)
+@Data
+@EqualsAndHashCode(callSuper = true)
+public class ServerConfigs extends AbstractPropertiesConfigs {
+    @MandatoryProperty
+    private String name;
+
+    // NOTE: test-purposes
+    public static ServerConfigs of(
+            String name
+    ) {
+        var instance = new ServerConfigs();
+        instance.name = name;
+        return instance;
+    }
+}